COM-1434-2 + REFACTO : change margin to padding + move AnswerTextArea
