+++ src/components/cards/AnswerTextArea.tsx
@@ -5,10 +5,10 @@ import {
   TextInput,
   View,
 } from 'react-native';
-import Shadow from '../../../components/style/Shadow';
-import { GREY, PINK, TRANSPARENT_PINK, WHITE } from '../../../styles/colors';
-import { BORDER_RADIUS, BORDER_WIDTH, MARGIN, PADDING, TEXT_AREA_HEIGHT } from '../../../styles/metrics';
-import { FIRA_SANS_MEDIUM, FIRA_SANS_REGULAR } from '../../../styles/fonts';
+import Shadow from '../style/Shadow';
+import { GREY, PINK, TRANSPARENT_PINK, WHITE } from '../../styles/colors';
+import { BORDER_RADIUS, BORDER_WIDTH, MARGIN, PADDING, TEXT_AREA_HEIGHT } from '../../styles/metrics';
+import { FIRA_SANS_MEDIUM, FIRA_SANS_REGULAR } from '../../styles/fonts';
 
 interface AnswerQuestionProps {
   answer: string,
