+++ src/screens/courses/cardTemplates/MultipleChoiceQuestionCard.tsx
@@ -8,7 +8,7 @@ import { getCard } from '../../../store/activities/selectors';
 import CardHeader from '../../../components/cards/CardHeader';
 import { FIRA_SANS_REGULAR } from '../../../styles/fonts';
 import { GREEN, GREY, ORANGE, PINK } from '../../../styles/colors';
-import { ABSOLUTE_BOTTOM_POSITION, INPUT_HEIGHT, MARGIN } from '../../../styles/metrics';
+import { ABSOLUTE_BOTTOM_POSITION, INPUT_HEIGHT, MARGIN, PADDING } from '../../../styles/metrics';
 import QuestionCardFooter from '../../../components/cards/QuestionCardFooter';
 import { MULTIPLE_CHOICE_QUESTION } from '../../../core/data/constants';
 import { navigate } from '../../../navigationRef';
@@ -116,7 +116,7 @@ const styles = (textColor: string, backgroundColor: string) => StyleSheet.create
     marginHorizontal: MARGIN.LG,
     flexGrow: 1,
     justifyContent: 'space-between',
-    marginBottom: MARGIN.LG,
+    paddingBottom: PADDING.XL,
   },
   informativeText: {
     ...FIRA_SANS_REGULAR.SM,
