+++ src/screens/courses/cardTemplates/OpenQuestion.tsx
@@ -10,7 +10,7 @@ import { GREY, PINK } from '../../../styles/colors';
 import { IS_LARGE_SCREEN, MARGIN } from '../../../styles/metrics';
 import QuestionCardFooter from '../../../components/cards/QuestionCardFooter';
 import { OPEN_QUESTION } from '../../../core/data/constants';
-import AnswerTextArea from './AnswerTextArea';
+import AnswerTextArea from '../../../components/cards/AnswerTextArea';
 import { QuestionnaireAnswerType } from '../../../types/store/ActivityStoreType';
 import Actions from '../../../store/activities/actions';
